flask_app/requirements.txt
black==25.12.0
blinker==1.9.0
click==8.3.1
flake8==7.3.0
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.2.1
pytokens==0.3.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.4